Split scripts into imports
