+++ assets/src/scripts/_plausible.js
@@ -1,0 +1,24 @@
+/**
+ * Plausible Analytics only on production
+ * Compat mode for IE11
+ */
+if (document.location.hostname === "reports.opensafely.org") {
+  const ua = window.navigator.userAgent;
+  const trident = ua.indexOf("Trident/");
+  const msie = ua.indexOf("MSIE ");
+
+  const script = document.createElement("script");
+  script.defer = true;
+  script.setAttribute("data-domain", "reports.opensafely.org");
+
+  // Serve legacy compat script to IE users
+  if (trident > 0 || msie > 0) {
+    script.id = "plausible";
+    script.src = "https://plausible.io/js/plausible.compat.js";
+  } else {
+    script.setAttribute("data-api", "/pa/api/event");
+    script.src = "/pa/js/script.js";
+  }
+
+  document.head.appendChild(script);
+}
